requirements.txt
selenium
beautifulsoup4
pandas
pyautogui
webdriver_manager
PyYAML
lxml
validate_email